netd: Allow devices to force-add directly-connected routes
Most sane devices have directly-connected routes properly
added by ConnectivityService, but a number of Samsung RIL
libraries seem to be doing special things, the end result
of which is IPv4 mobile data is non-functional due to lack
of a default route.

Affected devices have a log entry similar to the below:
  Error adding route 0.0.0.0/0 -> xxx.xxx.xxx.xxx rmnet0 to table 1002: Network is unreachable

This reverts, with guards, commit 92e8f96.

Change-Id: I6a653d19c1c72c136e78d8d841e87d3c166698ea
Signed-off-by: Joshua Primero <jprimero155@gmail.com>
Signed-off-by: HeroBuxx <herobuxx@gmail.com>